Refactor commands
- Each command is now in a separate file inside the "commands" package.
- Further, this change addresses the issue of erronous removal patches
  when a path with an array value was ignored.
- Also, it fixes a bug where Tailor would not find any item if both name
  and selector was given (oc returns a single item then instead of a
  list).
