+++ openshift/change.go
@@ -2,12 +2,8 @@ package openshift
 
 import (
 	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
 	"sort"
 
-	"github.com/opendevstack/tailor/cli"
 	"github.com/pmezard/go-difflib/difflib"
 )
 
@@ -31,12 +27,12 @@ type Change struct {
 	Action       string
 	Kind         string
 	Name         string
-	Patches      []*JsonPatch
+	Patches      []*jsonPatch
 	CurrentState string
 	DesiredState string
 }
 
-type JsonPatch struct {
+type jsonPatch struct {
 	Op    string      `json:"op"`
 	Path  string      `json:"path"`
 	Value interface{} `json:"value,omitempty"`
@@ -44,13 +40,6 @@ type JsonPatch struct {
 
 func (c *Change) ItemName() string {
 	return kindToShortMapping[c.Kind] + "/" + c.Name
-}
-
-func (c *Change) AddPatch(patch *JsonPatch) {
-	c.Patches = append(c.Patches, patch)
-	sort.Slice(c.Patches, func(i, j int) bool {
-		return c.Patches[i].Path < c.Patches[j].Path
-	})
 }
 
 func (c *Change) JsonPatches(pretty bool) string {
@@ -75,76 +64,9 @@ func (c *Change) Diff() string {
 	return text
 }
 
-func ocDelete(change *Change, compareOptions *cli.CompareOptions) error {
-	kind := change.Kind
-	name := change.Name
-	fmt.Printf("Deleting %s/%s ... ", kind, name)
-	args := []string{"delete", kind, name}
-	cmd := cli.ExecOcCmd(
-		args,
-		compareOptions.Namespace,
-		"", // empty as name and selector is not allowed
-	)
-	_, errBytes, err := cli.RunCmd(cmd)
-	if err == nil {
-		fmt.Println("done")
-	} else {
-		fmt.Println("failed")
-		return errors.New(string(errBytes))
-	}
-	return nil
-}
-
-func ocCreate(change *Change, compareOptions *cli.CompareOptions) error {
-	kind := change.Kind
-	name := change.Name
-	config := change.DesiredState
-	fmt.Printf("Creating %s/%s ... ", kind, name)
-	args := []string{"create", "-f", "-"}
-	cmd := cli.ExecOcCmd(
-		args,
-		compareOptions.Namespace,
-		compareOptions.Selector,
-	)
-	stdin, err := cmd.StdinPipe()
-	if err != nil {
-		return err
-	}
-	go func() {
-		defer stdin.Close()
-		io.WriteString(stdin, config)
-	}()
-	_, errBytes, err := cli.RunCmd(cmd)
-	if err == nil {
-		fmt.Println("done")
-	} else {
-		fmt.Println("failed")
-		return errors.New(string(errBytes))
-	}
-
-	return nil
-}
-
-func ocPatch(change *Change, compareOptions *cli.CompareOptions) error {
-	kind := change.Kind
-	name := change.Name
-
-	j := change.JsonPatches(false)
-
-	fmt.Printf("Patching %s/%s ... ", kind, name)
-
-	args := []string{"patch", kind + "/" + name, "--type=json", "--patch", j}
-	cmd := cli.ExecOcCmd(
-		args,
-		compareOptions.Namespace,
-		"", // empty as name and selector is not allowed
-	)
-	_, errBytes, err := cli.RunCmd(cmd)
-	if err == nil {
-		fmt.Println("done")
-	} else {
-		fmt.Println("failed")
-		return errors.New(string(errBytes))
-	}
-	return nil
+func (c *Change) addPatch(patch *jsonPatch) {
+	c.Patches = append(c.Patches, patch)
+	sort.Slice(c.Patches, func(i, j int) bool {
+		return c.Patches[i].Path < c.Patches[j].Path
+	})
 }
