+++ openshift/changeset_test.go
@@ -55,7 +55,9 @@ func TestAddDeleteOrder(t *testing.T) {
 func TestConfigNoop(t *testing.T) {
 
 	templateInput := []byte(
-		`apiVersion: v1
+		`kind: List
+metadata: {}
+apiVersion: v1
 items:
 - apiVersion: v1
   kind: PersistentVolumeClaim
@@ -70,14 +72,13 @@ items:
       requests:
         storage: 5Gi
     storageClassName: gp2
-  status: {}
-kind: List
-metadata: {}
-`)
+  status: {}`)
 
 	platformInput := []byte(
-		`apiVersion: v1
-items:
+		`kind: Template
+metadata: {}
+apiVersion: v1
+objects:
 - apiVersion: v1
   kind: PersistentVolumeClaim
   metadata:
@@ -96,31 +97,24 @@ items:
         storage: 5Gi
     storageClassName: gp2
     volumeName: pvc-2150713e-3e20-11e8-aa60-0aad3152d0e6
-  status: {}
-kind: List
-metadata: {}
-`)
+  status: {}`)
 
 	filter := &ResourceFilter{
 		Kinds: []string{"PersistentVolumeClaim"},
 	}
-	platformBasedList := &ResourceList{Filter: filter}
-	platformBasedList.CollectItemsFromPlatformList(platformInput)
-	templateBasedList := &ResourceList{Filter: filter}
-	templateBasedList.CollectItemsFromTemplateList(templateInput)
-	changeset, err := NewChangeset(platformBasedList, templateBasedList, false, []string{})
-	if err != nil {
-		t.Errorf(err.Error())
-	}
+	changeset := getChangeset(t, filter, platformInput, templateInput, false, []string{})
 	if !changeset.Blank() {
-		t.Errorf("Changeset is not blank, got %v", changeset.Update[0].JsonPatches(true))
+		//t.Errorf("Changeset is not blank, got %v", changeset.Update[0].JsonPatches(true))
+		t.Fail()
 	}
 }
 
 func TestConfigUpdate(t *testing.T) {
 
 	templateInput := []byte(
-		`apiVersion: v1
+		`kind: List
+metadata: {}
+apiVersion: v1
 items:
 - apiVersion: v1
   kind: PersistentVolumeClaim
@@ -135,14 +129,13 @@ items:
       requests:
         storage: 5Gi
     storageClassName: gp2
-  status: {}
-kind: List
-metadata: {}
-`)
+  status: {}`)
 
 	platformInput := []byte(
-		`apiVersion: v1
-items:
+		`kind: Template
+metadata: {}
+apiVersion: v1
+objects:
 - apiVersion: v1
   kind: PersistentVolumeClaim
   metadata:
@@ -157,30 +150,98 @@ items:
       requests:
         storage: 5Gi
     storageClassName: gp2
-  status: {}
-kind: List
-metadata: {}
-`)
+  status: {}`)
 
 	filter := &ResourceFilter{
 		Kinds: []string{"PersistentVolumeClaim"},
 	}
-	platformBasedList := &ResourceList{Filter: filter}
-	platformBasedList.CollectItemsFromPlatformList(platformInput)
-	templateBasedList := &ResourceList{Filter: filter}
-	templateBasedList.CollectItemsFromTemplateList(templateInput)
-	changeset, err := NewChangeset(platformBasedList, templateBasedList, false, []string{})
-	if err != nil {
-		t.Errorf(err.Error())
-	}
+	changeset := getChangeset(t, filter, platformInput, templateInput, false, []string{})
 	if len(changeset.Update) != 1 {
 		t.Errorf("Changeset.Update has %d items instead of 1", len(changeset.Update))
 	}
 }
 
+func TestConfigIgnoredPaths(t *testing.T) {
+	templateInput := []byte(
+		`kind: List
+apiVersion: v1
+items:
+- apiVersion: v1
+  kind: BuildConfig
+  metadata:
+    name: foo
+  spec:
+    failedBuildsHistoryLimit: 5
+    output:
+      to:
+        kind: ImageStreamTag
+        name: foo:latest
+    postCommit: {}
+    resources: {}
+    runPolicy: Serial
+    source:
+      binary: {}
+      type: Binary
+    strategy:
+      dockerStrategy: {}
+      type: Docker
+    successfulBuildsHistoryLimit: 5
+    triggers:
+    - generic:
+        secret: password
+      type: Generic`)
+
+	platformInput := []byte(
+		`kind: Template
+apiVersion: v1
+objects:
+- apiVersion: v1
+  kind: BuildConfig
+  metadata:
+    name: foo
+  spec:
+    failedBuildsHistoryLimit: 5
+    output:
+      to:
+        kind: ImageStreamTag
+        name: foo:abcdef
+      imageLabels:
+      - name: bar
+        value: baz
+    postCommit: {}
+    resources: {}
+    runPolicy: Serial
+    source:
+      binary: {}
+      type: Binary
+    strategy:
+      dockerStrategy: {}
+      type: Docker
+    successfulBuildsHistoryLimit: 5
+    triggers:
+    - generic:
+        secret: password
+      type: Generic`)
+
+	filter := &ResourceFilter{
+		Kinds: []string{"BuildConfig"},
+	}
+	changeset := getChangeset(t, filter, platformInput, templateInput, false, []string{"bc:/spec/output/to/name", "bc:/spec/output/imageLabels"})
+	actualUpdates := len(changeset.Update)
+	expectedUpdates := 0
+	if actualUpdates != expectedUpdates {
+		t.Errorf("Changeset.Update has %d items instead of %d", actualUpdates, expectedUpdates)
+		for i, u := range changeset.Update {
+			t.Errorf("Patchset Update#%d: %s", i, u.JsonPatches(true))
+		}
+	}
+}
+
 func TestConfigCreation(t *testing.T) {
 	templateInput := []byte(
-		`apiVersion: v1
+		`kind: List
+metadata: {}
+apiVersion: v1
 items:
 - apiVersion: v1
   kind: PersistentVolumeClaim
@@ -193,13 +254,13 @@ items:
       requests:
         storage: 5Gi
     storageClassName: gp2
-  status: {}
-kind: List
-metadata: {}`)
+  status: {}`)
 
 	platformInput := []byte(
-		`apiVersion: v1
-items:
+		`kind: Template
+metadata: {}
+apiVersion: v1
+objects:
 - apiVersion: v1
   kind: PersistentVolumeClaim
   metadata:
@@ -211,21 +272,12 @@ items:
       requests:
         storage: 5Gi
     storageClassName: gp2
-  status: {}
-kind: List
-metadata: {}`)
+  status: {}`)
 
 	filter := &ResourceFilter{
 		Kinds: []string{"PersistentVolumeClaim"},
 	}
-	platformBasedList := &ResourceList{Filter: filter}
-	platformBasedList.CollectItemsFromPlatformList(platformInput)
-	templateBasedList := &ResourceList{Filter: filter}
-	templateBasedList.CollectItemsFromTemplateList(templateInput)
-	changeset, err := NewChangeset(platformBasedList, templateBasedList, false, []string{})
-	if err != nil {
-		t.Errorf(err.Error())
-	}
+	changeset := getChangeset(t, filter, platformInput, templateInput, false, []string{})
 	if len(changeset.Create) != 1 {
 		t.Errorf("Changeset.Create is blank but should not be")
 	}
@@ -236,8 +288,10 @@ func TestConfigDeletion(t *testing.T) {
 	templateInput := []byte{}
 
 	platformInput := []byte(
-		`apiVersion: v1
-items:
+		`kind: Template
+metadata: {}
+apiVersion: v1
+objects:
 - apiVersion: v1
   kind: PersistentVolumeClaim
   metadata:
@@ -249,23 +303,29 @@ items:
       requests:
         storage: 5Gi
     storageClassName: gp2
-  status: {}
-kind: List
-metadata: {}
-`)
+  status: {}`)
 
 	filter := &ResourceFilter{
 		Kinds: []string{"PersistentVolumeClaim"},
 	}
-	platformBasedList := &ResourceList{Filter: filter}
-	platformBasedList.CollectItemsFromPlatformList(platformInput)
-	templateBasedList := &ResourceList{Filter: filter}
-	templateBasedList.CollectItemsFromTemplateList(templateInput)
-	changeset, err := NewChangeset(platformBasedList, templateBasedList, false, []string{})
-	if err != nil {
-		t.Errorf(err.Error())
-	}
+	changeset := getChangeset(t, filter, platformInput, templateInput, false, []string{})
 	if len(changeset.Delete) != 1 {
 		t.Errorf("Changeset.Delete is blank but should not be")
 	}
+}
+
+func getChangeset(t *testing.T, filter *ResourceFilter, platformInput, templateInput []byte, upsertOnly bool, ignoredPaths []string) *Changeset {
+	platformBasedList, err := NewPlatformBasedResourceList(filter, platformInput)
+	if err != nil {
+		t.Error("Could not create platform based list:", err)
+	}
+	templateBasedList, err := NewTemplateBasedResourceList(filter, templateInput)
+	if err != nil {
+		t.Error("Could not create template based list:", err)
+	}
+	changeset, err := NewChangeset(platformBasedList, templateBasedList, upsertOnly, ignoredPaths)
+	if err != nil {
+		t.Error("Could not create changeset:", err)
+	}
+	return changeset
 }
