+++ openshift/item.go
@@ -47,6 +47,26 @@ var (
 	platformModifiedFields = []string{
 		"/spec/template/spec/containers/[0-9]+/image$",
 	}
+
+	KindMapping = map[string]string{
+		"svc":                   "Service",
+		"service":               "Service",
+		"route":                 "Route",
+		"dc":                    "DeploymentConfig",
+		"deploymentconfig":      "DeploymentConfig",
+		"bc":                    "BuildConfig",
+		"buildconfig":           "BuildConfig",
+		"is":                    "ImageStream",
+		"imagestream":           "ImageStream",
+		"pvc":                   "PersistentVolumeClaim",
+		"persistentvolumeclaim": "PersistentVolumeClaim",
+		"template":              "Template",
+		"cm":                    "ConfigMap",
+		"configmap":             "ConfigMap",
+		"secret":                "Secret",
+		"rolebinding":           "RoleBinding",
+		"serviceaccount":        "ServiceAccount",
+	}
 )
 
 type ResourceItem struct {
@@ -80,7 +100,7 @@ func (templateItem *ResourceItem) ChangesFrom(platformItem *ResourceItem, extern
 		return nil, err
 	}
 
-	comparison := map[string]*JsonPatch{}
+	comparison := map[string]*jsonPatch{}
 	addedPaths := []string{}
 
 	for _, path := range templateItem.Paths {
@@ -98,7 +118,7 @@ func (templateItem *ResourceItem) ChangesFrom(platformItem *ResourceItem, extern
 			if templateItem.isImmutableField(path) {
 				return recreateChanges(templateItem, platformItem), nil
 			} else {
-				comparison[path] = &JsonPatch{Op: "add", Value: templateItemVal}
+				comparison[path] = &jsonPatch{Op: "add", Value: templateItemVal}
 				addedPaths = append(addedPaths, path)
 			}
 		} else {
@@ -106,21 +126,21 @@ func (templateItem *ResourceItem) ChangesFrom(platformItem *ResourceItem, extern
 			switch templateItemVal.(type) {
 			case []interface{}:
 				// slice content changed, continue ...
-				comparison[path] = &JsonPatch{Op: "noop"}
+				comparison[path] = &jsonPatch{Op: "noop"}
 			case []string:
 				// slice content changed, continue ...
-				comparison[path] = &JsonPatch{Op: "noop"}
+				comparison[path] = &jsonPatch{Op: "noop"}
 			case map[string]interface{}:
 				// map content changed, continue
-				comparison[path] = &JsonPatch{Op: "noop"}
+				comparison[path] = &jsonPatch{Op: "noop"}
 			default:
 				if templateItemVal == platformItemVal {
-					comparison[path] = &JsonPatch{Op: "noop"}
+					comparison[path] = &jsonPatch{Op: "noop"}
 				} else {
 					if templateItem.isImmutableField(path) {
 						return recreateChanges(templateItem, platformItem), nil
 					} else {
-						comparison[path] = &JsonPatch{Op: "replace", Value: templateItemVal}
+						comparison[path] = &jsonPatch{Op: "replace", Value: templateItemVal}
 					}
 				}
 			}
@@ -136,7 +156,7 @@ func (templateItem *ResourceItem) ChangesFrom(platformItem *ResourceItem, extern
 				continue
 			}
 			// Pointer exist only in platformItem
-			comparison[path] = &JsonPatch{Op: "remove"}
+			comparison[path] = &jsonPatch{Op: "remove"}
 			deletedPaths = append(deletedPaths, path)
 		}
 	}
@@ -144,15 +164,16 @@ func (templateItem *ResourceItem) ChangesFrom(platformItem *ResourceItem, extern
 	c := &Change{
 		Kind:         templateItem.Kind,
 		Name:         templateItem.Name,
-		Patches:      []*JsonPatch{},
+		Patches:      []*jsonPatch{},
 		CurrentState: platformItem.YamlConfig(),
 		DesiredState: templateItem.YamlConfig(),
 	}
 
 	for path, patch := range comparison {
 		if patch.Op != "noop" {
+			cli.DebugMsg("add path", path)
 			patch.Path = path
-			c.AddPatch(patch)
+			c.addPatch(patch)
 		}
 	}
 
@@ -407,17 +428,26 @@ func (templateItem *ResourceItem) prepareForComparisonWithPlatformItem(platformI
 		platformItemVal, _, err := pathPointer.Get(platformItem.Config)
 		if err != nil {
 			cli.DebugMsg("No such path", path, "in platform item", platformItem.FullName())
-		}
-		_, err = pathPointer.Set(templateItem.Config, platformItemVal)
-		if err != nil {
-			cli.DebugMsg(
-				"Could not set",
-				path,
-				"to",
-				platformItemVal.(string),
-				"in template item",
-				templateItem.FullName(),
-			)
+		} else {
+			_, err = pathPointer.Set(templateItem.Config, platformItemVal)
+			if err != nil {
+				cli.DebugMsg(fmt.Sprintf(
+					"Could not set %s to %v in template item %s",
+					path,
+					platformItemVal,
+					templateItem.FullName(),
+				))
+			} else {
+				// Add ignored path and its subpaths to the paths slice
+				// of the template item.
+				templateItem.Paths = append(templateItem.Paths, path)
+				switch vv := platformItemVal.(type) {
+				case []interface{}:
+					templateItem.walkArray(vv, path)
+				case map[string]interface{}:
+					templateItem.walkMap(vv, path)
+				}
+			}
 		}
 	}
