+++ openshift/item_test.go
@@ -117,12 +117,12 @@ func TestChangesFromAnnotationFields(t *testing.T) {
 	}
 	actualPatchOne := changes[0].Patches[0]
 	actualPatchTwo := changes[0].Patches[1]
-	expectedPatchOne := &JsonPatch{
+	expectedPatchOne := &jsonPatch{
 		Op:    "add",
 		Path:  "/metadata/annotations/foo",
 		Value: "bar",
 	}
-	expectedPatchTwo := &JsonPatch{
+	expectedPatchTwo := &jsonPatch{
 		Op:    "add",
 		Path:  "/metadata/annotations/managed-annotations.tailor.opendevstack.org",
 		Value: "foo",
@@ -141,8 +141,8 @@ func TestChangesFromAnnotationFields(t *testing.T) {
 	if err != nil {
 		t.Errorf(err.Error())
 	}
-	var actualPatch *JsonPatch
-	var expectedPatch *JsonPatch
+	var actualPatch *jsonPatch
+	var expectedPatch *jsonPatch
 	if len(changes) > 1 || changes[0].Action != "Noop" {
 		actualPatch = changes[0].Patches[0]
 		t.Errorf("Platform and template should have no drift, got %v", actualPatch)
@@ -159,7 +159,7 @@ func TestChangesFromAnnotationFields(t *testing.T) {
 		t.Errorf("Platform and template should have drift")
 	}
 	actualPatch = changes[0].Patches[0]
-	expectedPatch = &JsonPatch{
+	expectedPatch = &jsonPatch{
 		Op:    "add",
 		Path:  "/metadata/annotations/managed-annotations.tailor.opendevstack.org",
 		Value: "foo",
@@ -179,7 +179,7 @@ func TestChangesFromAnnotationFields(t *testing.T) {
 		t.Errorf("Platform and template should have drift")
 	}
 	actualPatch = changes[0].Patches[0]
-	expectedPatch = &JsonPatch{
+	expectedPatch = &jsonPatch{
 		Op:    "replace",
 		Path:  "/metadata/annotations/foo",
 		Value: "baz",
@@ -204,7 +204,7 @@ func TestChangesFromAnnotationFields(t *testing.T) {
 		t.Errorf("Platform and template should have drift")
 	}
 	actualPatch = changes[0].Patches[0]
-	expectedPatch = &JsonPatch{
+	expectedPatch = &jsonPatch{
 		Op:    "replace",
 		Path:  "/metadata/annotations/foo",
 		Value: "baz",
@@ -223,7 +223,7 @@ func TestChangesFromAnnotationFields(t *testing.T) {
 		t.Errorf("Platform and template should have drift")
 	}
 	actualPatch = changes[0].Patches[0]
-	expectedPatch = &JsonPatch{
+	expectedPatch = &jsonPatch{
 		Op:   "remove",
 		Path: "/metadata/annotations/foo",
 	}
