+++ openshift/template_test.go
